requirements.txt
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==4.0.0
dj-database-url==0.5.0
psycopg2-binary==2.9.9
django-environ==0.9.0
Pillow==10.2.0
drf-yasg==1.21.7 